Move data entry structure to global types+utils
