+++ frontend/src/features/data_entry/types/types.ts
@@ -1,7 +1,6 @@
 import { Dispatch } from "react";
 
 import { AnyApiError } from "@/api/ApiResult";
-import { TranslationPath } from "@/i18n/i18n.types";
 import {
   ClaimDataEntryResponse,
   DataEntryStatus,
@@ -9,63 +8,9 @@ import {
   PollingStationResults,
   ValidationResults,
 } from "@/types/generated/openapi";
-import { FormSectionId, PollingStationResultsPath } from "@/types/types";
+import { DataEntryStructure, FormSectionId, SectionValues } from "@/types/types";
 
 import { ValidationResultSet } from "../utils/ValidationResults";
-
-// Data Entry Section Types
-export interface HeadingSubsection {
-  type: "heading";
-  title: TranslationPath;
-}
-
-export interface MessageSubsection {
-  type: "message";
-  message: TranslationPath;
-  className?: string;
-}
-
-export interface RadioSubsectionOption {
-  value: string;
-  label: TranslationPath;
-  autoFocusInput?: boolean;
-}
-
-export interface RadioSubsection {
-  type: "radio";
-  error: TranslationPath;
-  path: PollingStationResultsPath;
-  options: RadioSubsectionOption[];
-  valueType?: "string" | "boolean";
-}
-
-export interface InputGridSubsectionRow {
-  code?: string;
-  path: PollingStationResultsPath;
-  title?: string;
-  isTotal?: boolean;
-  isListTotal?: boolean;
-  addSeparator?: boolean;
-  autoFocusInput?: boolean;
-}
-
-export interface InputGridSubsection {
-  type: "inputGrid";
-  headers: [TranslationPath, TranslationPath, TranslationPath];
-  zebra?: boolean;
-  rows: InputGridSubsectionRow[];
-}
-
-export type DataEntrySubsection = HeadingSubsection | MessageSubsection | RadioSubsection | InputGridSubsection;
-
-export interface DataEntrySection {
-  id: FormSectionId;
-  title: string;
-  short_title: string;
-  subsections: DataEntrySubsection[];
-}
-
-export type DataEntryStructure = DataEntrySection[];
 
 export interface DataEntryState {
   // state from providers
@@ -154,8 +99,6 @@ export interface SubmitCurrentFormOptions {
   continueToNextSection?: boolean;
   showAcceptErrorsAndWarnings?: boolean;
 }
-
-export type SectionValues = Record<string, string>;
 
 //store unvalidated data
 export type TemporaryCache = {
